separate components and state
