+++ src/components/IngredientsList.jsx
@@ -1,0 +1,25 @@
+function IngredientsList({ ingredients, toggleRecipeShown }) {
+  const ingredientsListItems = ingredients.map((ingredient) => {
+    return <li key={ingredient}>{ingredient}</li>;
+  });
+
+  return (
+    <section aria-labelledby="ingredients-header">
+      <h2 id="ingredients-header">Ingredients on hand:</h2>
+      <ul className="ingredients-list" aria-live="polite">
+        {ingredientsListItems}
+      </ul>
+      {ingredients.length >= 4 && (
+        <div className="recipe-results">
+          <div>
+            <h3>Ready for a recipe?</h3>
+            <p>Generate a recipe from your list of ingredients.</p>
+          </div>
+          <button onClick={toggleRecipeShown}>Get a recipe</button>
+        </div>
+      )}
+    </section>
+  );
+}
+
+export default IngredientsList;
