+++ src/components/Recipe.jsx
@@ -1,0 +1,5 @@
+function Recipe() {
+  return <p>This is the recipe</p>;
+}
+
+export default Recipe;
